fix: BatchQuery

+++ pkg/metrics/interface/interface.go
@@ -9,7 +9,7 @@ import (
 type QueryFormItem struct {
 	Start int64  `json:"start" binding:"required"` // 开始时间
 	End   int64  `json:"end" binding:"required"`   // 结束时间
-	Step  int64  `json:"step" binding:"required"`  // 步长
+	Step  uint   `json:"step" binding:"required"`  // 步长
 	Query string `json:"query" binding:"required"` // 查询语句
 }
